Rename asyncStateSlice files to useAsyncState2 to center on hook
