+++ applications/browser-extension/src/utils/interview/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { asyncStateReducer } from "@/utils/interview/asyncStateSlice";
+import { asyncStateReducer } from "@/utils/interview/useAsyncState2";
 
 const store = configureStore({
   reducer: {
